extracted repo abstractions into folders in domain, fixed namespaces
