+++ src/Modules/InventoryManagement/DigiDent.InventoryManagement.Domain/Employees/ReadModels/EmployeeDetails.cs
@@ -1,7 +1,7 @@
 ﻿using DigiDent.InventoryManagement.Domain.Actions.ReadModels;
 using DigiDent.InventoryManagement.Domain.Requests.ReadModels;
 
-namespace DigiDent.InventoryManagement.Domain.Employees;
+namespace DigiDent.InventoryManagement.Domain.Employees.ReadModels;
 
 public record EmployeeDetails: EmployeeSummary
 {
