+++ src/Modules/InventoryManagement/DigiDent.InventoryManagement.Domain/Employees/ReadModels/EmployeeSummary.cs
@@ -1,6 +1,6 @@
 ﻿using DigiDent.Shared.Kernel.Pagination;
 
-namespace DigiDent.InventoryManagement.Domain.Employees;
+namespace DigiDent.InventoryManagement.Domain.Employees.ReadModels;
 
 public record EmployeeSummary: IFilterable, ISortable
 {
