+++ src/Modules/InventoryManagement/DigiDent.InventoryManagement.Domain/Employees/Repositories/IEmployeesCommandsRepository.cs
@@ -1,4 +1,4 @@
-﻿namespace DigiDent.InventoryManagement.Domain.Employees;
+﻿namespace DigiDent.InventoryManagement.Domain.Employees.Repositories;
 
 public interface IEmployeesCommandsRepository
 {
