+++ src/Modules/InventoryManagement/DigiDent.InventoryManagement.Domain/Requests/Repositories/IRequestsCommandsRepository.cs
@@ -1,6 +1,6 @@
 ﻿using DigiDent.InventoryManagement.Domain.Requests.ValueObjects;
 
-namespace DigiDent.InventoryManagement.Domain.Requests;
+namespace DigiDent.InventoryManagement.Domain.Requests.Repositories;
 
 public interface IRequestsCommandsRepository
 {
