+++ src/Modules/InventoryManagement/DigiDent.InventoryManagement.Persistence/Actions/InventoryActionsQueriesRepository.cs
@@ -1,8 +1,10 @@
 ﻿using Dapper;
 using DigiDent.InventoryManagement.Domain.Actions;
 using DigiDent.InventoryManagement.Domain.Actions.ReadModels;
+using DigiDent.InventoryManagement.Domain.Actions.Repositories;
 using DigiDent.InventoryManagement.Domain.Actions.ValueObjects;
 using DigiDent.InventoryManagement.Domain.Employees;
+using DigiDent.InventoryManagement.Domain.Employees.ReadModels;
 using DigiDent.InventoryManagement.Domain.Items.ReadModels;
 using DigiDent.InventoryManagement.Domain.Items.ValueObjects;
 using DigiDent.InventoryManagement.Persistence.Constants;
