+++ src/Modules/InventoryManagement/DigiDent.InventoryManagement.Persistence/Items/InventoryItemsQueriesRepository.cs
@@ -1,8 +1,10 @@
 ﻿using Dapper;
 using DigiDent.InventoryManagement.Domain.Actions.ReadModels;
 using DigiDent.InventoryManagement.Domain.Employees;
+using DigiDent.InventoryManagement.Domain.Employees.ReadModels;
 using DigiDent.InventoryManagement.Domain.Items;
 using DigiDent.InventoryManagement.Domain.Items.ReadModels;
+using DigiDent.InventoryManagement.Domain.Items.Repositories;
 using DigiDent.InventoryManagement.Domain.Items.ValueObjects;
 using DigiDent.InventoryManagement.Domain.Requests.ReadModels;
 using DigiDent.InventoryManagement.Persistence.Constants;
